KB-12-4 rename
