+++ src/main/kotlin/lessson_12/lesson_12_4.kt
@@ -1,5 +1,8 @@
-package lessson_12_4
+package lessson_12
 
+fun main(){
+    val seventhDayLife = KeepWeather4(12, 13, atmosphericPressure = 660)
+}
 
 class KeepWeather4(
     var temperatureDay: Int,
@@ -21,6 +24,4 @@ class KeepWeather4(
           """.trimMargin()
         )
     }
-
-
 }
