--- src/main/kotlin/lessson_12_4/main.kt
@@ -1,5 +1,0 @@
-package lessson_12_4
-
-fun main(){
-    val seventhDayLife = KeepWeather4(12, 13, atmosphericPressure = 660)
-}
